+++ src/lib.rs
@@ -1,6 +1,12 @@
 // Copyright (C) 2019-2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
+#![deny(broken_intra_doc_links, missing_docs)]
+
+//! A crate providing a replacement #[[macro@test]] attribute that
+//! initializes logging and/or tracing infrastructure before running
+//! tests.
+
 extern crate proc_macro;
 
 use proc_macro::TokenStream;