+++ tests/mod.rs
@@ -1,6 +1,8 @@
 // Copyright (C) 2019-2021 Daniel Mueller <deso@posteo.net>
 // SPDX-License-Identifier: (Apache-2.0 OR MIT)
 
+#![allow(clippy::eq_op)]
+
 use tokio::runtime::Builder;
 
 use tracing::debug;