+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-Unreleased
-----------
+0.2.9
+-----
 - Added support for inner `#[test]` attribute arguments
 - Added CI checks for auto generated code documentation
 - Bumped minimum supported Rust version to `1.49`